Remove `ome_zarr` subpackage, and move its content to main package
