+++ examples/01_cardio_tiny_dataset/run_workflow_partial_channel_inputs.py
@@ -16,7 +16,7 @@ from pathlib import Path
 
 from devtools import debug
 
-from fractal_tasks_core.ome_zarr.channels import OmeroChannel
+from fractal_tasks_core.channels import OmeroChannel
 from fractal_tasks_core.tasks.create_ome_zarr import create_ome_zarr
 from fractal_tasks_core.tasks.yokogawa_to_ome_zarr import yokogawa_to_ome_zarr
