+++ examples/04_cardio_tiny_dataset_napari_worfklows/run_example_workflow_6.py
@@ -16,8 +16,8 @@ from pathlib import Path
 
 from devtools import debug
 
-from fractal_tasks_core.ome_zarr.channels import OmeroChannel
-from fractal_tasks_core.ome_zarr.channels import Window
+from fractal_tasks_core.channels import OmeroChannel
+from fractal_tasks_core.channels import Window
 from fractal_tasks_core.tasks.create_ome_zarr import create_ome_zarr
 from fractal_tasks_core.tasks.napari_workflows_wrapper import (
     napari_workflows_wrapper,
