+++ examples/08_subset_of_data/run_example_single_channel_few_Z_planes.py
@@ -2,8 +2,8 @@ import os
 
 from devtools import debug
 
-from fractal_tasks_core.ome_zarr.channels import OmeroChannel
-from fractal_tasks_core.ome_zarr.channels import Window
+from fractal_tasks_core.channels import OmeroChannel
+from fractal_tasks_core.channels import Window
 from fractal_tasks_core.tasks.copy_ome_zarr import copy_ome_zarr
 from fractal_tasks_core.tasks.create_ome_zarr import create_ome_zarr
 from fractal_tasks_core.tasks.maximum_intensity_projection import (
