+++ fractal_tasks_core/masked_loading.py
