+++ fractal_tasks_core/tasks/apply_registration_to_image.py
@@ -27,8 +27,7 @@ import zarr
 from pydantic.decorator import validate_arguments
 
 from fractal_tasks_core.ngff import load_NgffImageMeta
-from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
-from fractal_tasks_core.ome_zarr.utils import get_table_path_dict
+from fractal_tasks_core.pyramids import build_pyramid
 from fractal_tasks_core.roi import (
     convert_indices_to_regions,
 )
@@ -38,6 +37,7 @@ from fractal_tasks_core.roi import (
 from fractal_tasks_core.roi import is_standard_roi_table
 from fractal_tasks_core.roi import load_region
 from fractal_tasks_core.tables import write_table
+from fractal_tasks_core.utils import get_table_path_dict
 
 logger = logging.getLogger(__name__)
