+++ fractal_tasks_core/tasks/cellpose_segmentation.py
@@ -31,15 +31,14 @@ from cellpose import models
 from pydantic.decorator import validate_arguments
 
 import fractal_tasks_core
+from fractal_tasks_core.channels import ChannelInputModel
+from fractal_tasks_core.channels import ChannelNotFoundError
+from fractal_tasks_core.channels import get_channel_from_image_zarr
+from fractal_tasks_core.channels import OmeroChannel
 from fractal_tasks_core.labels import prepare_label_group
+from fractal_tasks_core.masked_loading import masked_loading_wrapper
 from fractal_tasks_core.ngff import load_NgffImageMeta
-from fractal_tasks_core.ome_zarr.channels import ChannelInputModel
-from fractal_tasks_core.ome_zarr.channels import ChannelNotFoundError
-from fractal_tasks_core.ome_zarr.channels import get_channel_from_image_zarr
-from fractal_tasks_core.ome_zarr.channels import OmeroChannel
-from fractal_tasks_core.ome_zarr.masked_loading import masked_loading_wrapper
-from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
-from fractal_tasks_core.ome_zarr.utils import rescale_datasets
+from fractal_tasks_core.pyramids import build_pyramid
 from fractal_tasks_core.roi import (
     array_to_bounding_box_table,
 )
@@ -55,6 +54,7 @@ from fractal_tasks_core.roi import get_overlapping_pairs_3D
 from fractal_tasks_core.roi import is_ROI_table_valid
 from fractal_tasks_core.roi import load_region
 from fractal_tasks_core.tables import write_table
+from fractal_tasks_core.utils import rescale_datasets
 
 logger = logging.getLogger(__name__)
