+++ fractal_tasks_core/tasks/maximum_intensity_projection.py
@@ -22,7 +22,7 @@ from pydantic.decorator import validate_arguments
 from zarr.errors import ContainsArrayError
 
 from fractal_tasks_core.ngff import load_NgffImageMeta
-from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
+from fractal_tasks_core.pyramids import build_pyramid
 from fractal_tasks_core.zarr_utils import OverwriteNotAllowedError
 
 logger = logging.getLogger(__name__)
