+++ fractal_tasks_core/tasks/yokogawa_to_ome_zarr.py
@@ -28,16 +28,16 @@ from fractal_tasks_core.cellvoyager.filenames import (
     glob_with_multiple_patterns,
 )
 from fractal_tasks_core.cellvoyager.filenames import parse_filename
+from fractal_tasks_core.channels import get_omero_channel_list
+from fractal_tasks_core.channels import OmeroChannel
 from fractal_tasks_core.ngff import load_NgffImageMeta
-from fractal_tasks_core.ome_zarr.channels import get_omero_channel_list
-from fractal_tasks_core.ome_zarr.channels import OmeroChannel
-from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
-from fractal_tasks_core.ome_zarr.utils import (
-    get_parameters_from_metadata,
-)
+from fractal_tasks_core.pyramids import build_pyramid
 from fractal_tasks_core.roi import check_valid_ROI_indices
 from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
+)
+from fractal_tasks_core.utils import (
+    get_parameters_from_metadata,
 )
 from fractal_tasks_core.zarr_utils import OverwriteNotAllowedError
