+++ tests/data/fake_package_for_schema_experiments/my_package/tasks/my_task.py
@@ -1,5 +1,5 @@
 from ..lib_custom_models import CustomModel
-from fractal_tasks_core.ome_zarr.channels import OmeroChannel
+from fractal_tasks_core.channels import OmeroChannel
 
 
 def my_task(x: int, y: OmeroChannel, z: CustomModel):
